Code cleanup - move GraphPathToTripPlanConverter to package org.opentripplanner.routing.algorithm.mapping.
